+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changes
 2.0.1 (2017-xx-xx)
 ------------------
 
-- Added `request_info` to response object.
+- Added `request_info` to response object and `ClientResponseError`.
 
 2.0.0 (2017-03-20)
 ------------------